Move `_fetch` function into it's own service
